fastapi/requirements.txt
fastapi
uvicorn
aiolinebot
line-bot-sdk
python-dotenv
python-multipart
numpy
opencv-python
python-gyazo